requirements.txt
selenium>=4.16.0
webdriver-manager>=4.0.1
python-dotenv>=1.0.0
fastapi>=0.109.0
uvicorn>=0.27.0
sqlalchemy>=2.0.25
pydantic>=2.5.0
requests>=2.31.0
apscheduler>=3.10.4
